Rename Schema::Subset -> Schema::Visibility
