+++ spec/graphql/logger_spec.rb
@@ -96,8 +96,8 @@ describe "Logger" do
 
     it "logs about hidden interfaces with no implementations" do
       res = LoggerTest::CustomLoggerSchema.execute("{ node(id: \"5\") { id } }", context: { skip_types_migration_error: true })
-      if GraphQL::Schema.use_schema_subset?
-        assert_nil res["data"]["node"], "Schema::Subset doesn't warn in this case -- it doesn't check possible types because it doesn't have to"
+      if GraphQL::Schema.use_schema_visibility?
+        assert_nil res["data"]["node"], "Schema::Visibility::Subset doesn't warn in this case -- it doesn't check possible types because it doesn't have to"
       else
         assert_equal ["Field 'node' doesn't exist on type 'Query'"], res["errors"].map { |err| err["message"] }
         assert_includes LoggerTest::CustomLoggerSchema::LOG_STRING.string, "Interface `Node` hidden because it has no visible implementers"
@@ -110,8 +110,8 @@ describe "Logger" do
         res = LoggerTest::DefaultLoggerSchema.execute("{ node(id: \"5\") { id } }", context: { skip_types_migration_error: true })
       end
 
-      if GraphQL::Schema.use_schema_subset?
-        assert_nil res["data"]["node"], "Schema::Subset doesn't warn in this case -- it doesn't check possible types because it doesn't have to"
+      if GraphQL::Schema.use_schema_visibility?
+        assert_nil res["data"]["node"], "Schema::Visibility::Subset doesn't warn in this case -- it doesn't check possible types because it doesn't have to"
       else
         assert_equal ["Field 'node' doesn't exist on type 'Query'"], res["errors"].map { |err| err["message"] }
       end
