+++ spec/graphql/schema/addition_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 describe GraphQL::Schema::Addition do
   it "handles duplicate types with cycles" do
     duplicate_types_schema = Class.new(GraphQL::Schema)
-    duplicate_types_schema.use_schema_subset = false
+    duplicate_types_schema.use_schema_visibility = false
     duplicate_types = 2.times.map {
       Class.new(GraphQL::Schema::Object) do
         graphql_name "Thing"
