+++ spec/integration/rails/graphql/schema_spec.rb
@@ -16,7 +16,7 @@ describe GraphQL::Schema do
 
   describe "#union_memberships" do
     it "returns a list of unions that include the type" do
-      skip("Not implemented for Subset") if GraphQL::Schema.use_schema_subset?
+      skip("Not implemented for Subset") if GraphQL::Schema.use_schema_visibility?
       assert_equal [schema.types["Animal"], schema.types["AnimalAsCow"]], schema.union_memberships(schema.types["Cow"])
     end
   end
@@ -43,7 +43,7 @@ describe GraphQL::Schema do
 
   describe "#references_to" do
     it "returns a list of Field and Arguments of that type" do
-      skip "Not implemented when using Subset" if GraphQL::Schema.use_schema_subset?
+      skip "Not implemented when using Subset" if GraphQL::Schema.use_schema_visibility?
       cow_field = schema.get_field("Query", "cow")
       cow_t = schema.get_type("Cow")
       assert_equal [cow_field], schema.references_to(cow_t)
