+++ spec/spec_helper.rb
@@ -19,8 +19,8 @@ end
 if ENV["GRAPHQL_REJECT_NUMBERS_FOLLOWED_BY_NAMES"]
   puts "Opting into GraphQL.reject_numbers_followed_by_names"
   GraphQL.reject_numbers_followed_by_names = true
-  puts "Opting into GraphQL::Schema::Subset"
-  GraphQL::Schema.use_schema_subset = true
+  puts "Opting into GraphQL::Schema::Visibility::Subset"
+  GraphQL::Schema.use(GraphQL::Schema::Visibility)
   GraphQL::Schema.use(GraphQL::Schema::TypesMigration)
 end
